+++ openeo/rest/connection.py
@@ -25,7 +25,7 @@ from openeo.rest.auth.auth import NullAuth, BearerAuth
 from openeo.rest.auth.config import RefreshTokenStore, AuthConfig
 from openeo.rest.auth.oidc import OidcClientCredentialsAuthenticator, OidcAuthCodePkceAuthenticator, \
     OidcClientInfo, OidcAuthenticator, OidcRefreshTokenAuthenticator, OidcResourceOwnerPasswordAuthenticator, \
-    OidcDeviceAuthenticator, OidcProviderInfo
+    OidcDeviceAuthenticator, OidcProviderInfo, OidcException
 from openeo.rest.datacube import DataCube
 from openeo.rest.imagecollectionclient import ImageCollectionClient
 from openeo.rest.job import RESTJob
@@ -277,35 +277,6 @@ class Connection(RestApiConnection):
             self.auth = BearerAuth(bearer=resp["access_token"])
         return self
 
-    @deprecated("Use :py:meth:`authenticate_oidc_authorization_code` or something similar.", version="0.3.5")
-    def authenticate_OIDC(
-            self, client_id: str,
-            provider_id: str = None,
-            webbrowser_open=None,
-            timeout=120,
-            server_address: Tuple[str, int] = None
-    ) -> 'Connection':
-        """
-        Authenticates a user to the backend using OpenID Connect.
-
-        :param client_id: Client id to use for OpenID Connect authentication
-        :param webbrowser_open: optional handler for the initial OAuth authentication request
-            (opens a webbrowser by default)
-        :param timeout: number of seconds after which to abort the authentication procedure
-        :param server_address: optional tuple (hostname, port_number) to serve the OAuth redirect callback on
-        """
-        # TODO: option to increase log level temporarily?
-        provider_id, provider = self._get_oidc_provider(provider_id)
-
-        client_info = OidcClientInfo(client_id=client_id, provider=provider)
-        authenticator = OidcAuthCodePkceAuthenticator(
-            client_info=client_info,
-            webbrowser_open=webbrowser_open,
-            timeout=timeout,
-            server_address=server_address,
-        )
-        return self._authenticate_oidc(authenticator, provider_id=provider_id)
-
     def _get_oidc_provider(self, provider_id: Union[str, None] = None) -> Tuple[str, OidcProviderInfo]:
         """
         Get OpenID Connect discovery URL for given provider_id
@@ -320,18 +291,34 @@ class Connection(RestApiConnection):
             _log.info("Found OIDC providers: {p}".format(p=list(providers.keys())))
             if provider_id:
                 if provider_id not in providers:
-                    raise OpenEoClientException("Requested provider {r!r} not available. Should be one of {p}.".format(
-                        r=provider_id, p=list(providers.keys()))
+                    raise OpenEoClientException(
+                        "Requested OIDC provider {r!r} not available. Should be one of {p}.".format(
+                            r=provider_id, p=list(providers.keys())
+                        )
                     )
                 provider = providers[provider_id]
             elif len(providers) == 1:
-                # No provider id given, but there is only one anyway: we can handle that.
                 provider_id, provider = providers.popitem()
+                _log.info("No OIDC provider given, but only one available: {p!r}. Use that one.".format(
+                    p=provider_id
+                ))
             else:
-                raise OpenEoClientException("No provider_id given. Available: {p!r}.".format(
-                    p=list(providers.keys()))
-                )
-            provider = OidcProviderInfo(issuer=provider["issuer"], scopes=provider.get("scopes"))
+                # Check if there is a single provider in the config to use.
+                backend = self._orig_url
+                provider_configs = self._get_auth_config().get_oidc_provider_configs(backend=backend)
+                intersection = set(provider_configs.keys()).intersection(providers.keys())
+                if len(intersection) == 1:
+                    provider_id = intersection.pop()
+                    provider = providers[provider_id]
+                    _log.info(
+                        "No OIDC provider id given, but only one in config (backend {b!r}): {p!r}."
+                        " Use that one.".format(b=backend, p=provider_id)
+                    )
+                else:
+                    raise OpenEoClientException("No OIDC provider id given. Pick one from: {p!r}.".format(
+                        p=list(providers.keys()))
+                    )
+            provider = OidcProviderInfo.from_dict(provider)
         else:
             # Per spec: '/credentials/oidc' will redirect to  OpenID Connect discovery document
             provider = OidcProviderInfo(discovery_url=self.build_url('/credentials/oidc'))
@@ -347,7 +334,7 @@ class Connection(RestApiConnection):
         :param provider_id: id of OIDC provider as specified by backend (/credentials/oidc).
             Can be None if there is just one provider.
 
-        :return: (client_id, client_secret)
+        :return: OIDC provider id and client info
         """
         provider_id, provider = self._get_oidc_provider(provider_id)
 
@@ -355,9 +342,18 @@ class Connection(RestApiConnection):
             client_id, client_secret = self._get_auth_config().get_oidc_client_configs(
                 backend=self._orig_url, provider_id=provider_id
             )
-            _log.info("Using client_id {c!r} from config (provider {p!r})".format(c=client_id, p=provider_id))
-            if client_id is None:
-                raise OpenEoClientException("No client ID found.")
+            if client_id:
+                _log.info("Using client_id {c!r} from config (provider {p!r})".format(c=client_id, p=provider_id))
+        if client_id is None:
+            # TODO: This "default_client" feature is still experimental in openEO API. See Open-EO/openeo-api#366
+            # Try "default_client" from backend's provider info.
+            client_id = provider.get_default_client_id()
+            if client_id:
+                _log.info("Using default client_id {c!r} from OIDC provider {p!r} info.".format(
+                    c=client_id, p=provider_id
+                ))
+        if client_id is None:
+            raise OpenEoClientException("No client ID found.")
 
         client_info = OidcClientInfo(client_id=client_id, client_secret=client_secret, provider=provider)
 
@@ -489,6 +485,45 @@ class Connection(RestApiConnection):
         )
         authenticator = OidcDeviceAuthenticator(client_info=client_info, use_pkce=use_pkce, **kwargs)
         return self._authenticate_oidc(authenticator, provider_id=provider_id, store_refresh_token=store_refresh_token)
+
+    def authenticate_oidc(
+            self,
+            provider_id: str = None,
+            client_id: Union[str, None] = None, client_secret: Union[str, None] = None,
+            store_refresh_token: bool = True
+    ):
+        """
+        Do OpenID Connect authentication, first trying refresh tokens and falling back on device code flow.
+        """
+        provider_id, client_info = self._get_oidc_provider_and_client_info(
+            provider_id=provider_id, client_id=client_id, client_secret=client_secret
+        )
+
+        # Try refresh token first.
+        refresh_token = self._get_refresh_token_store().get_refresh_token(
+            issuer=client_info.provider.issuer,
+            client_id=client_info.client_id
+        )
+        if refresh_token:
+            try:
+                _log.info("Found refresh token: trying refresh token based authentication.")
+                authenticator = OidcRefreshTokenAuthenticator(client_info=client_info, refresh_token=refresh_token)
+                con = self._authenticate_oidc(
+                    authenticator, provider_id=provider_id, store_refresh_token=store_refresh_token
+                )
+                # TODO: pluggable/jupyter-aware display function?
+                print("Authenticated using refresh token.")
+                return con
+            except OidcException as e:
+                _log.info("Refresh token based authentication failed: {e}.".format(e=e))
+
+        # Fall back on device code flow
+        # TODO: make it possible to do other fallback flows too?
+        _log.info("Trying device code flow.")
+        authenticator = OidcDeviceAuthenticator(client_info=client_info)
+        con = self._authenticate_oidc(authenticator, provider_id=provider_id, store_refresh_token=store_refresh_token)
+        print("Authenticated using device code flow.")
+        return con
 
     def describe_account(self) -> str:
         """
@@ -902,7 +937,7 @@ def connect(url, auth_type: str = None, auth_options: dict = {}, session: reques
         >>> # For basic authentication
         >>> conn = connect(url).authenticate_basic(username="john", password="foo")
         >>> # For OpenID Connect authentication
-        >>> conn = connect(url).authenticate_OIDC(client_id="myclient")
+        >>> conn = connect(url).authenticate_oidc(client_id="myclient")
 
     :param url: The http url of an OpenEO endpoint.
     :param auth_type: Which authentication to use: None, "basic" or "oidc" (for OpenID Connect)
@@ -917,7 +952,7 @@ def connect(url, auth_type: str = None, auth_options: dict = {}, session: reques
     elif auth_type == "basic":
         connection.authenticate_basic(**auth_options)
     elif auth_type in {"oidc", "openid"}:
-        connection.authenticate_OIDC(**auth_options)
+        connection.authenticate_oidc(**auth_options)
     else:
         raise ValueError("Unknown auth type {a!r}".format(a=auth_type))
     return connection